--- a/engine/src/main/coffee/engine/core/abstractagentset.coffee
+++ b/engine/src/main/coffee/engine/core/abstractagentset.coffee
@@ -80,7 +80,7 @@ module.exports =
 
     # () => Iterator
     iterator: ->
-      new Iterator(@_agentArr)
+      new Iterator(@_agentArr[..])
 
     # (() => Number) => AbstractAgentSet[T]
     maxesBy: (f) ->