--- a/engine/src/main/coffee/engine/core/linkset.coffee
+++ b/engine/src/main/coffee/engine/core/linkset.coffee
@@ -13,7 +13,11 @@ module.exports =
 
     # () => Iterator[T]
     iterator: ->
-      new DeadSkippingIterator(@_unwrap(@_agents))
+      new DeadSkippingIterator(@_unwrap(@_agents, true))
+
+    # () => Iterator[T]
+    _unsafeIterator: ->
+      new DeadSkippingIterator(@_unwrap(@_agents, false))
 
     # I know, I know, this is insane, right?  "Why would you do this?!", you demand.  I don't blame you.  I don't like
     # it, either.  But, look... we have a problem on our hands.  Special agentsets are a thing.  They can be stored
@@ -24,5 +28,5 @@ module.exports =
     # array, or we need to use a good sorting structure that represents the data as an array under the hood, and then
     # we can pass around that array.  Passing thunks seems to be the better option to me.  --JAB (9/7/14)
     # [T] @ ((() => Array[T])|Array[T]) => Array[T]
-    _unwrap: (agents) ->
-      if JSType(agents).isFunction() then agents() else agents[..]
+    _unwrap: (agents, copy) ->
+      if JSType(agents).isFunction() then agents() else if copy then agents[..] else agents