… resources/test/dumps/Rock Paper Scissors.js